# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RestoraPro Restaurant Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: tableId
          in: query
          schema:
            type: integer
        - name: waiterId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/items:
    post:
      operationId: addOrderItems
      tags: [orders]
      summary: Add items to order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddOrderItemsBody"
      responses:
        "201":
          description: Items added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
  /order-items/{id}/status:
    patch:
      operationId: updateOrderItemStatus
      tags: [orders]
      summary: Update order item status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderItemStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        tableId:
          type: integer
        tableNumber:
          type: integer
        waiterId:
          type: integer
        waiterName:
          type: string
        status:
          type: string
          enum: [open, bill_requested, paid, cancelled]
        subtotal:
          type: number
        tax:
          type: number
        serviceCharge:
          type: number
        total:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, tableId, waiterId, status, subtotal, total]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        department:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, accepted, preparing, ready, served]
        sentAt:
          type: string
          nullable: true
        readyAt:
          type: string
          nullable: true
      required: [id, orderId, menuItemId, quantity, unitPrice, status]
    OrderWithItems:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
          required: [items]
    CreateOrderBody:
      type: object
      properties:
        tableId:
          type: integer
        waiterId:
          type: integer
        notes:
          type: string
      required: [tableId, waiterId]
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
    AddOrderItemsBody:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: integer
              quantity:
                type: integer
              notes:
                type: string
            required: [menuItemId, quantity]
      required: [items]
    UpdateOrderItemStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, accepted, preparing, ready, served]
      required: [status]
...